code pour les objets

+++ WorldGenerator/include/Arbre.h
@@ -6,7 +6,8 @@
 #include "Ellipse.h"
 #include "Object.h"
 
-// Composite Design pour eviter l'heritage diament mais on aurrait pu utiliser de l'heritage vituel
+// Composite Design pour eviter l'heritage diamant mais on aurrait pu utiliser de l'heritage vituel
+// FAUX ! On a un vecteur de polygon et d'elipse donc l'heritage virtuel ne peut pas creer l'objet
 // lien vers heritage virtuel
 // https://cpp.developpez.com/faq/cpp/?page=L-heritage
 // https://www.ljll.math.upmc.fr/courscpp/Sections/Sect08-E3.html
@@ -14,7 +15,7 @@ class Arbre : public Object
 {
     public:
         Arbre(){};
-        Arbre(Coords const & lacoord,int taille);
+        Arbre(Coords const & lacoord,double taille);
         virtual ~Arbre(){};
         virtual void medessiner();
